move fixture specific helpers
