+++ src/util/__tests__/helpers.test.ts
@@ -1,7 +1,6 @@
 import { AddressTuples } from "../../models/types/scene-to-fixture-assignment.ts";
 import {
   merge16BitValues,
-  handleChannelValues,
   convertDmxValueToPercent,
   percentageToColorTemperature,
   percentageToIntensityLevel,
@@ -33,90 +32,6 @@ describe("Merge 16bit values", () => {
       [7, 255],
       [8, 255],
     ]);
-  });
-
-  describe("handleChannelValues properly merges 16bit channels, and 8bit channel with profile name", () => {
-    test("it correctly handles 16bit channels", () => {
-      const profileChannels1 = {
-        1: "Dimmer",
-        2: "Dimmer fine",
-      };
-
-      const values1: AddressTuples = [
-        [1, 255],
-        [2, 255],
-        [3, 128],
-        [4, 128],
-      ];
-
-      const channelPairs16Bit1: AddressTuples = [
-        [1, 2],
-        [3, 4],
-      ];
-
-      const { result: result1 } = handleChannelValues(
-        profileChannels1,
-        values1,
-        channelPairs16Bit1,
-        true,
-        () => true,
-      );
-
-      expect(result1).toStrictEqual({ Dimmer: 65535 });
-
-      const profileChannels2 = {
-        3: "Color Temp",
-        4: "Color Temp fine",
-      };
-
-      const values2: AddressTuples = [
-        [1, 255],
-        [2, 255],
-        [3, 128],
-        [4, 128],
-      ];
-
-      const channelPairs16Bit2: AddressTuples = [
-        [1, 2],
-        [3, 4],
-      ];
-
-      const { result: result2 } = handleChannelValues(
-        profileChannels2,
-        values2,
-        channelPairs16Bit2,
-        true,
-        () => true,
-      );
-      expect(result2).toStrictEqual({ "Color Temp": 32896 });
-    });
-
-    test("it handles 16bit values that contain 0", () => {
-      const profileChannels = {
-        1: "Dimmer",
-        2: "Dimmer fine",
-        3: "Color Temp",
-        4: "Color Temp fine",
-      };
-      const values: AddressTuples = [
-        [1, 255],
-        [2, 0],
-      ];
-      const channelPairs16Bit = [
-        [1, 2],
-        [3, 4],
-      ];
-
-      const { result } = handleChannelValues(
-        profileChannels,
-        values,
-        channelPairs16Bit,
-        true,
-        () => true,
-      );
-
-      expect(result).toStrictEqual({ Dimmer: 65280 });
-    });
   });
 
   describe("Convert DMX Value to Percent", () => {
